moved common constants to common.ts
