+++ airstack-modules/modules/airstack/nft-marketplace/nft.ts
@@ -15,7 +15,8 @@ import {
     AirMeta
 } from "../../../generated/schema";
 
-import { AIR_NFT_SALE_ENTITY_ID, AIR_META_ID, BIGINT_ONE, SUBGRAPH_SCHEMA_VERSION, SUBGRAPH_VERSION, SUBGRAPH_NAME, SUBGRAPH_SLUG, processNetwork } from "./utils";
+import { AIR_NFT_SALE_ENTITY_ID } from "./utils";
+import { AIR_META_ID, BIGINT_ONE, SUBGRAPH_SCHEMA_VERSION, SUBGRAPH_VERSION, SUBGRAPH_NAME, SUBGRAPH_SLUG, processNetwork } from "../common";
 
 export namespace nft {
     export function trackNFTSaleTransactions(
